pyproject.toml
[project]
name = "ai-agent-backend"
version = "1.0.0"
description = "Python 後端 API 程式，整合 LangChain 和 LangGraph 開發的 AI Agent"
authors = [
    {name = "AI Agent Developer", email = "developer@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.30.0,<1.0.0",
    "pydantic>=2.7.2,<3.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-openai>=0.1.0,<1.0.0",
    "langchain-google-genai>=2.0.0,<3.0.0",
    "langgraph>=0.0.30,<1.0.0",
    "openai>=1.6.1,<2.0.0",
    "google-generativeai>=0.8.0,<1.0.0",
    "python-dotenv==1.0.0",
    "aiohttp>=3.9.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "httpx>=0.27.0,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agents"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto" 